### 2020-10-27 #2785 @whedon
[PRE REVIEW]: PySensors: A Python Package for Sparse Sensor Placement
**Submitting author:** @briandesilva (<a href="http://orcid.org/0000-0003-0944-900X">Brian de Silva</a>)
**Repository:** <a href="https://github.com/dynamicslab/pysensors/" target ="_blank">https://github.com/dynamicslab/pysensors/</a>
**Version:** v.0.3.0
