### 2020-10-27 766b3fb
Add missing DOIs
