+++ docs/JOSS/paper.bib
@@ -17,7 +17,8 @@
   number={3},
   pages={63--86},
   year={2018},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/mcs.2018.2810460}
 }
 
 @article{brunton2016sparse,
@@ -49,7 +50,8 @@
   volume={681},
   number={241-260},
   year={2011},
-  publisher={Cambridge University Press, The Edinburgh Building, Cambridge CB 2 2 RU~…}
+  publisher={Cambridge University Press, The Edinburgh Building, Cambridge CB 2 2 RU~…},
+  doi={10.1017/jfm.2011.195}
 }
 
 @article{krause2008near,
@@ -81,7 +83,8 @@
   number={2},
   pages={208--226},
   year={2006},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.2514/6.2004-2415}
 }
 
 @article{brunton2016pnas,
@@ -111,7 +114,8 @@
   title = {A unified sparse optimization framework to learn parsimonious physics-informed models from data},
   author = {Champion, Kathleen and Zheng, Peng and Aravkin, Aleksandr Y. and Brunton, Steven L. and Kutz, J. Nathan},
   journal = {arXiv preprint arXiv:1906.10612},
-  year = {2019}
+  year = {2019},
+  doi={10.1109/access.2020.3023625}
 }
 
 @article {Schmidt81,
@@ -335,7 +339,8 @@
   Author = {Loiseau, Jean-Christophe},
   Journal = {arXiv preprint arXiv:1911.07920},
   Title = {Data-driven modeling of the chaotic thermal convection in an annular thermosyphon},
-  Year = {2019}
+  Year = {2019},
+  doi={10.1007/s00162-020-00536-w}
 }
 
 @article{Reinbold2020pre,
@@ -367,14 +372,16 @@
   Pages = {160--173},
   Title = {Efficient sensor placement for ocean measurements using low-dimensional concepts},
   Volume = {27},
-  Year = {2009}}
+  Year = {2009},
+  doi={10.1016/j.ocemod.2009.01.001}}
 
 @article{Dhingra2014cdc,
   Author = {N. K. Dhingra AND M. R. Jovanovic AND Z.-Q. Luo},
   Journal = {53rd IEEE Conference on Decision and Control},
   Pages = {4039--4044},
   Title = {An {ADMM} algorithm for optimal sensor and actuator selection},
-  Year = {2014}}
+  Year = {2014},
+  doi={10.1109/cdc.2014.7040017}}
 
 @article{Munz2014ieeetac,
   Author = {U. Munz AND M. Pfister AND P. Wolfrum},
@@ -391,7 +398,8 @@
   Pages = {366--381},
   Title = {Local flow characterization using bioinspired sensory information},
   Volume = {818},
-  Year = {2017}}
+  Year = {2017},
+  doi={10.1017/jfm.2017.137}}
 
 @article{Mohren2018pnas,
   Author = {Mohren, Thomas L and Daniel, Thomas L and Brunton, Steven L and Brunton, Bingni W},
@@ -400,7 +408,8 @@
   Pages = {10564--10569},
   Title = {Neural-inspired sensors enable sparse, efficient classification of spatiotemporal data},
   Volume = {115},
-  Year = {2018}}
+  Year = {2018},
+  doi={10.1073/pnas.1808909115}}
 
 @article{Manohar2018arxivB,
   Author = {K. Manohar and J. N. Kutz and S. L. Brunton},
@@ -414,7 +423,8 @@
   Pages = {87--95},
   Title = {Predicting shim gaps in aircraft assembly with machine learning and sparse sensing},
   Volume = {48},
[…]
