### 2020-10-28 #2785 @arfon
> Failed to discover a `Statement of need` section in paper

:wave: @briandesilva - could you please make sure your paper has a clear statement of need for a general, non-specialist audience?
