### 2020-10-28 fa8327b
Add statement of need section to paper
