+++ docs/JOSS/paper.bib
@@ -752,3 +752,10 @@
   Volume = {39},
   Year = {2017},
   doi={10.1137/16m1081270}}
+
+@article{klise2017sensor,
+  title={Sensor placement optimization using Chama},
+  author={Klise, Katherine A and Nicholson, Bethany and Laird, Carl D},
+  journal={Number SAND2017-11472. Albuquerque, NM: Sandia National Laboratories},
+  year={2017}
+}
