### 2020-11-02 b17c156
Fix typo in docstring
