### 2020-11-05 #2785 @pdebuyl
Hi @briandesilva thank you for submitting to JOSS. I will look for reviewers shortly.

In the meantime, I have a few comments
- The second paragraph of the article is a laundry list of citations. JOSS papers should be written for a non-specialist audience and include "a list of key references".
- The list of author of the fourth reference is lacking full names.
- Please also check that the DOI are properly formatted for all references.
- You do not cite scikit-learn despite having prepared your library for interacting with it. This should be fixed.
