### 2020-11-05 #2785 @pdebuyl
@stsievert , @jordanperr can you review the article "PySensors: A Python Package for Sparse Sensor Placement" for JOSS ?
