### 2020-11-05 #2785 @pdebuyl
Thank you for the reply @jordanperr . We have documentation for authors / reviewers / editors at https://joss.readthedocs.io/en/latest/index.html

The high-level view is this: when we have two reviewers, I start the review. There will be a new github issue with the checklist and a link to the instructions. The issue page serves as the communication channel for all involved and I'll be available to reply to questions. You can have a look at accepted papers reviews if you wish https://github.com/openjournals/joss-reviews/issues?page=5&q=+label%3Aaccepted

For now, you should just wait for the review to start :-)
