### 2020-11-06 d73a321
Remove extra citations, add missing DOIS, reformat some citations
