+++ docs/JOSS/paper.bib
@@ -6,7 +6,8 @@
   number={7},
   pages={2642--2656},
   year={2018},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/JSEN.2018.2887044},
 }
 
 @article{manohar2018data,
@@ -29,7 +30,8 @@
   number={5},
   pages={2099--2122},
   year={2016},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/15M1036713}
 }
 
 @article{summers2015submodularity,
@@ -418,13 +420,15 @@
   Year = {2018}}
 
 @article{Manohar2018jms,
-  Author = {K. Manohar AND T. Hogan AND J. Buttrick AND A. G. Banerjee AND J. N. Kutz AND S. L. Brunton},
-  Journal = {Journal of Manufacturing Systems},
-  Pages = {87--95},
-  Title = {Predicting shim gaps in aircraft assembly with machine learning and sparse sensing},
-  Volume = {48},
-  Year = {2018},
-  doi={10.1016/j.jmsy.2018.01.011}}
+  title={Predicting shim gaps in aircraft assembly with machine learning and sparse sensing},
+  author={Manohar, Krithika and Hogan, Thomas and Buttrick, Jim and Banerjee, Ashis G and Kutz, J Nathan and Brunton, Steven L},
+  journal={Journal of manufacturing systems},
+  volume={48},
+  pages={87--95},
+  year={2018},
+  publisher={Elsevier},
+  doi={10.1016/j.jmsy.2018.01.011}
+}
 
 @article{Zare2018arxiv,
   Author = {Armin Zare and Neil K. Dhingra and Mihailo R. Jovanovi{\'c} and Tryphon T. Georgiou},
@@ -439,13 +443,15 @@
   Title = {Sparse reduced-order modeling: sensor-based dynamics to full-state estimation},
   Volume = {844},
   Year = {2018},
-  doi={10.1017/jfm.2018.147}}
+  DOI={10.1017/jfm.2018.147}}
 
 @book{Brunton2019book,
-  Author = {S. L. Brunton and J. N. Kutz},
+  Author = {Steven L. Brunton and J. Nathan Kutz},
   Publisher = {Cambridge University Press},
   Title = {Data-Driven Science and Engineering: Machine Learning, Dynamical Systems, and Control},
-  Year = {2019}}
+  Year = {2019},
+  DOI = {10.1017/9781108380690}
+  }
 
 @article{bieker2020deep,
   Author = {Bieker, Katharina and Peitz, Sebastian and Brunton, Steven L and Kutz, J Nathan and Dellnitz, Michael},
@@ -540,7 +546,9 @@
   Pages = {1289--1306},
   Title = {Compressed sensing},
   Volume = {52},
-  Year = {2006}}
+  Year = {2006},
+  DOI = {10.1109/TIT.2006.871582},
+  }
 
 @article{Candes2006picm,
   Author = {E. J. Cand\`es},
@@ -584,7 +592,9 @@
   Pages = {118--120},
   Title = {Compressive sensing},
   Volume = {24},
-  Year = {2007}}
+  Year = {2007},
+  DOI = {10.1109/MSP.2007.4286571},
+  }
 
 @article{Candes2008ieeespm,
   Author = {E. J. Cand\`es AND M. B. Wakin},
@@ -713,7 +723,9 @@
   Pages = {A631--A648},
   Title = {A New Selection Operator for the Discrete Empirical Interpolation Method---Improved A Priori Error Bound and Extensions},
   Volume = {38},
-  Year = {2016}}
+  Year = {2016},
+  DOI = {10.1137/15M1019271}
+}
 
 @article{DiLorenzo2016ieee,
[…]
