### 2020-11-06 #2785 @briandesilva
@pdebuyl, I just pushed the changes you requested. Please let me know if the DOIs aren't formatted the way you had in mind.
