### 2020-11-06 #2785 @pdebuyl
Thank you @briandesilva , apart from Karniadakis 2009 and Tao the references are fine and my other comments addressed.
