### 2020-11-07 b8aa2af
Fix JOSS paper citations
