+++ docs/JOSS/paper.bib
@@ -369,7 +369,7 @@
 }
 
 @article{Yildirim:2009,
-  Author = {B. Yildirim AND C. Chryssostomidis AND G. E. Karniadakis},
+  Author = {Yildirim, B and Chryssostomidis, C and Karniadakis, GE},
   Journal = {Ocean Modelling},
   Pages = {160--173},
   Title = {Efficient sensor placement for ocean measurements using low-dimensional concepts},
@@ -470,7 +470,7 @@
   Year = {1956}}
 
 @article{Candes2006cpam,
-  Author = {E. J. Cand\`es AND J. Romberg AND T. Tao},
+  Author = {Cand\`es, Emmanuel J and Romberg, Justin K and Tao, Terence},
   Journal = {Communications in Pure and Applied Mathematics},
   Number = {1207--1223},
   Title = {Stable signal recovery from incomplete and inaccurate measurements},
