### 2020-11-07 #2785 @briandesilva
> apart from Karniadakis 2009 and Tao the references are fine and my other comments addressed.
Good catch—fixed.
