### 2020-11-09 #2785 @pdebuyl
@tuelwer can you review the article "PySensors: A Python Package for Sparse Sensor Placement" for JOSS ?
