### 2020-11-10 dynamicslab/pysensors#10 @tuelwer
Code example in README.rst is outdated
While [reviewing](https://github.com/openjournals/joss-reviews/issues/2828) your toolbox for JOSS I found out that the reconstruction example in the [README.rst](https://github.com/dynamicslab/pysensors/blob/master/README.rst) is outdated:

```text
>>> model = pysensors.Sensorselector(n_sensors=10)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
AttributeError: module 'pysensors' has no attribute 'Sensorselector'
```

I think `pysensors.Sensorselector` has been renamed to `pysensors.reconstruction.SSPOR()`.
