### 2020-11-10 ff71379
Update SensorSelector --> SSPOR in README
