### 2020-11-10 c0b7ef7
Fix lingering DOI issues
