+++ docs/JOSS/paper.bib
@@ -769,7 +769,8 @@
   title={Sensor placement optimization using Chama},
   author={Klise, Katherine A and Nicholson, Bethany and Laird, Carl D},
   journal={Number SAND2017-11472. Albuquerque, NM: Sandia National Laboratories},
-  year={2017}
+  year={2017},
+  doi={10.2172/1405271},
 }
 
 @article{scikit-learn,
@@ -782,5 +783,4 @@
  volume={12},
  pages={2825--2830},
  year={2011},
- doi={10.5555/1953048.2078195},
 }
