### 2020-11-10 dynamicslab/pysensors#10 @briandesilva
Thanks for pointing this out! You're exactly right—we renamed `SensorSelector` in the code, but missed it here. I just pushed a fix.
