### 2020-11-10 dynamicslab/pysensors#10 @tuelwer
I still have trouble running the code of the reconstruction example. For example `data` is never used and `method.selected_sensors` should probably be `model.get_selected_sensors()`. Please correct me if I'm wrong.
