### 2020-11-11 af68e29
Fix mistakes in README
