### 2020-11-11 dynamicslab/pysensors#10 @briandesilva
I'm terribly sorry—I should have looked more closely at the example. Everything should be fixed now. This time I've tested the commands to make sure they run properly. Please let me know if you have any more trouble.

Note that both `model.get_selected_sensors()` and `model.selected_sensors` are allowed (though `method` was incorrect).
