### 2020-11-11 dynamicslab/pysensors#10 @tuelwer
Thanks! Seems to work now!
