### 2020-11-11 #2828 @tuelwer
@pdebuyl @briandesilva  The authors propose the Python library PySensors which allows efficient data-driven sensor placement. The toolbox provides methods for reconstruction and classification based on the data of the sensors. The main contribution of the paper is the implementation of the sparse sensor placement optimization algorithm for reconstruction (SSPOR) and classification (SSPOC). Besides this, the toolbox provides useful bindings to existing methods. The API is designed in the style of scikit-learn which makes the toolbox very easy to use!

The paper is well-written and gives a good overview over related work. However, I would have liked a more formal definition of the problem setting that is solved by the PySensors toolbox. The references seem to be complete. The scikit-learn JMLR paper has indeed no DOI stated on the website of the journal. There is a missing DOI for Chama which whedon found.

The toolbox itself is well-documented. I especially like the example notebooks which are nicely written and give a good overview over the features of the toolbox. I was able to reproduce the results of the examples locally on my notebook as well as on binder. However, per default, binder is missing matplotlib, seaborn and pandas. Can this be configured? Installation of the toolbox on MacOS, Debian and Ubuntu through `pip install .` and `pip install python-sensors` worked without problems. All tests passed with some warnings. There were some minor problems with the reconstruction example in the `README.rst` which are described in [this issue](https://github.com/dynamicslab/pysensors/issues/10).

Regarding the community guide: I cannot see a clear statement on how third parties could seek support.

I congratulate the authors to their great toolbox! Overall, I would recommend to accept this submission into JOSS.

Minor remarks:
- The last reference in the `README.rst` is printed in bold. Is this done on purpose?
- In example notebook `pysensors_overview.ipynb` two different UserWarnings are thrown.
- Some example notebooks require additional packages like seaborn or netCDF4. It would be nice to have a `requirements-examples.txt` to install all those packages in a single run.
- Figure in the reconstruction example of the docs is missing.
- "the number of possible placements grows combinatorially" (first paragraph in paper): "combinatorially" -> "exponentially"
