### 2020-11-11 #2828 @briandesilva
Thanks, @tuelwer! This is all very useful feedback. We'll work to update the paper and package accordingly.
