### 2020-11-13 28767ac
Move notebook requirements to separate file
