+++ requirements-examples.txt
@@ -1,0 +1,6 @@
+jupyter
+matplotlib
+netCDF4
+notebook
+pandas
+seaborn
