### 2020-11-13 622f001
Slight wording change in paper
