### 2020-11-13 #2828 @briandesilva
@tuelwer, could you clarify which figure you're referring to here?
> Figure in the reconstruction example of the docs is missing.

I believe I've addressed the other issues you raised, apart from the binder issue and the "more formal definition of the problem setting that is solved by the PySensors toolbox." I experimented with different solutions for specifying binder dependencies, but didn't come up with anything that worked. I'll have to look into both items further.
