### 2020-11-13 #2828 @tuelwer
@briandesilva What I meant was [this example](https://python-sensors.readthedocs.io/en/latest/index.html). I was indeed somewhat imprecise, sorry for this!

![missing_image](https://user-images.githubusercontent.com/7251921/99049573-4aca6680-2597-11eb-851f-d4ab19e5d75d.png)
