### 2020-11-18 #2828 @briandesilva
Okay I've fixed the binder issue and the missing image. @tuelwer, with regards to your comment
>I would have liked a more formal definition of the problem setting that is solved by the PySensors toolbox

there are different objective functions that are approximately optimized by PySensors classes depending on the problem type. We had hesitated to include them in our JOSS submission based on our understanding that the paper is meant to be aimed at a general audience. What do you think about the idea of adding the objective functions to their respective Jupyter notebooks?
