### 2020-11-21 #2828 @briandesilva
@kmanohar, do you want to take a stab at adding a brief description of the problem setting to [the paper](https://github.com/dynamicslab/pysensors/blob/master/docs/JOSS/paper.md)?
