### 2020-11-23 #2828 @jordanperr
Review for PySensors: A Python Package for Sparse Sensor Placement

Pysensors is a Python implementation of the SSPOC (pysensors.classification.SSPOC) and SSPOR (pysensors.reconstruction.SSPOR) methods to perform optimization of sensor placement. These methods are published in (Brunton, 2016) which is cited in both the documentation and in the paper. The repository under review contains documentation (hosted on readthedocs and in Readme markdown files) and worked out examples in the form of Jupyter notebooks. The code is thoughtfully organized, with two optimizers (CCQR and QR) that extend scikit-learn's BaseEstimator, and two utility functions that wrap MultiTaskLasso and OrthogonalMatchingPursuit from scipy. I was able to install the software, run the tests, and execute some of the example notebooks.

The paper is well written and fits within the scope of the journal. I believe the code represents a significant contribution as a user friendly implementation of the SSPOR and SSPOC algorithms. I would have liked to see citations of academic work that use the PySensor package. Such citations would provide evidence of the impact of this software in an academic field.

Overall, **I recommend acceptance for publication** after these minor comments are reviewed and implemented at the author's discretion:

**Paper:**

- Citation would be useful for the claim “In general, choosing the globally optimal placement… is an intractable computation”.
- The notion of "sensors" and "optimization of sensor placement" could be considered as domain specific jargon for a truly general audience. The authors may consider defining these terms so the general reader has a more concrete picture of the problem being solved here.
- A note on the difference between “detection” and “reconstruction and classification” tasks would be helpful.
- In "statement of need," it may be construed from the text that Chama is the only other software to solve the sparse sensor placement problem. Quickly searching "sensor placement" on Github yields 77 repositories, including SSPOR_pub repo from K. Manohar (an author of this paper), which appears to implement the SSPOR algorithm in MATLAB. Some other repos include "sustainability-lab/polire", and "Chandrayee/sensor-placement". How do these codes fit into the story?
	- The authors could mention how optimal sensor placement is currently solved by each researchers with custom scripts on top of more general statistical learning and optimization packages, such as scikit-learn or scipy. A brief overview of the more general techniques such as SVD, POD, or DMD would be beneficial and help explain why choosing PySensors is superior to a custom script.
- Citations of academic work that use PySensors could help prove the impact of the software. I do believe that researchers in my peer group would use this software (and I will certainly send it around to them!)

**Documentation:**

- At time of review, binaries for the scikit-learn dependency for Python 3.9.0, Mac OSX, are not available. The installation thus fails trying to compile scikit-learn from source. I specified python 3.8.5 and the installation worked as expected.
- The readme assumes familiarity with Python and pip. This assumption is reasonable, but may not be valid for all academic users.
	- Consider including a high level description of the dependencies in the Readme, such as “Python 3.6-3.8 with Mac or Linux and pip”. Not immediately clear if this software supports Windows.
	- May not be clear for all users what the next step is after "pip install ." A simple sentence with a link to the examples section would help.
- There is something weird going on with the TOC on the ReadTheDocs page. I see "pysensors.version module" and "Module contents" as subheadings under "API" - which doesn't seem right. Otherwise, the RTD page looks very good.

**Examples:**

- The example notebooks have a lot of warnings. To pick on one: classification.ipynb (as checked in to github) has the following warning after cell 19:
```
/home/brian/Documents/Dropbox (uwamath)/Brian/Research/PySensors/pySensors/venv/lib/python3.6/site-packages/sklearn/linear_model/_coordinate_descent.py:1914: UserWarning: Coordinate descent with l1_reg=0 may lead to unexpected results and is discouraged.
  check_random_state(self.random_state), random)
```

**Code:**

- Tests pass with 36 warnings. They all look like the aforementioned "UserWarning" caused by improper parameters / small data, which is to be expected with tests. These warnings could be caught by the tests and ignored.
- Performance is not mentioned in the documentation. Would be useful to know the time and space complexity of these algorithms.
- Excellent code style and software architecture. I believe the implementation is exemplary, although I am not familiar with the pattern of naming each module with an underscore prefix and then renaming them in the `__init__` scripts.
