### 2020-11-30 ebbf4e5
added description of problem statement to paper.md
