### 2020-12-01 ce3c24c
Replace html math with LaTeX
