### 2020-12-01 b346973
Flesh out installation instructions in README
