### 2020-12-06 9b5e3f5
Remove version module from docs
