### 2020-12-06 9a5bafe
Suppress warnings in tests
