### 2020-12-06 22490fe
Fix call to deprecated sphinx function:
