### 2020-12-06 4a7063b
Better control of warnings. Remove all warnings from examples
