### 2020-12-07 #2828 @tuelwer
@briandesilva, @kmanohar thank you for adding the description of the problem setting. Some minor remarks: the index i is not defined. Perhaps you want to add `for i=1...n` or something similar. Also it would be nice to state the shape of y, C and x and to make clear that y usually has less entries than x.
