### 2020-12-11 9611b9d
updated references and problem description in paper.md
