+++ docs/JOSS/paper.bib
@@ -1,3 +1,33 @@
+@article{ko1995exact,
+  title={An exact algorithm for maximum entropy sampling},
+  author={Ko, Chun-Wa and Lee, Jon and Queyranne, Maurice},
+  journal={Operations Research},
+  volume={43},
+  number={4},
+  pages={684--691},
+  year={1995},
+  publisher={INFORMS}
+}
+
+@inproceedings{narayanan2020toolkit,
+  title={A toolkit for spatial interpolation and sensor placement},
+  author={Narayanan, S Deepak and Patel, Zeel B and Agnihotri, Apoorv and Batra, Nipun},
+  booktitle={Proceedings of the 18th Conference on Embedded Networked Sensor Systems},
+  pages={653--654},
+  year={2020}
+}
+
+@article{manohar2019optimized,
+  title={Optimized sampling for multiscale dynamics},
+  author={Manohar, Krithika and Kaiser, Eurika and Brunton, Steven L and Kutz, J Nathan},
+  journal={Multiscale Modeling \& Simulation},
+  volume={17},
+  number={1},
+  pages={117--136},
+  year={2019},
+  publisher={SIAM}
+}
+
 @article{clark2018greedy,
   title={Greedy sensor placement with cost constraints},
   author={Clark, Emily and Askham, Travis and Brunton, Steven L and Kutz, J Nathan},
