### 2020-12-12 d2481e5
updated problem description in paper.md
