### 2020-12-14 8ba0d4f
Minor statement of need updates
