### 2020-12-14 #2828 @briandesilva
@jordanperr @tuelwer, thanks again for all your feedback. We have pushed a new version of the paper in which we have attempted to address your comments.

> Some other repos include "sustainability-lab/polire", and "Chandrayee/sensor-placement". How do these codes fit into the story?

We were unable to find any documentation for the package "Chandrayee/sensor-placement" and hence did not include it in the paper, but based on the code, it appears to optimize the same Gaussian process criteria as "sustainability-lab/polire".

> Citations of academic work that use PySensors could help prove the impact of the software.

As far as we know, PySensors has not been used in any published academic work as of yet.
