### 2020-12-14 #2828 @pdebuyl
Thank you @tuelwer for the review!

@briandesilva some of @jordanperr comments remain not entirely answered, as I understand it. A note on the performance is missing, for instance.

Also, could you state that the code is untested under windows, if that is well the case?

Some papers lack a DOI, such as 10.1137/17M1162366 (even if whedon did not list it, you should check that everything that has a DOI is listed with the DOI in the paper).
