+++ docs/JOSS/paper.bib
@@ -6,7 +6,8 @@
   number={4},
   pages={684--691},
   year={1995},
-  publisher={INFORMS}
+  publisher={INFORMS},
+  doi={10.1287/opre.43.4.684},
 }
 
 @inproceedings{narayanan2020toolkit,
@@ -25,7 +26,8 @@
   number={1},
   pages={117--136},
   year={2019},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/17M1162366},
 }
 
 @article{clark2018greedy,
