### 2020-12-16 7ad44df
Add note about windows
