### 2020-12-16 #2828 @briandesilva
@pdebuyl, I have fixed the DOIs and added a line to the README about lack of testing for windows machines.

I will have to go back and look over @jordanperr's comments to be double-check for anything I've missed. Though I'll note that I previously added performance discussions to various sections of the documentation.
