### 2020-12-18 #2828 @pdebuyl
Hi @briandesilva the note on windows is fine, thank you. Regarding performance, I only looked at the paper, it is fine in the doc sorry.
