### 2020-12-20 6df8c5e
Add explanation of sensor
