### 2020-12-20 #2828 @briandesilva
Hi @pdebuyl, I've looked back over @jordanperr's feedback and saw that the following comment may not have been completely addressed:

> The notion of "sensors" and "optimization of sensor placement" could be considered as domain specific jargon for a truly general audience. The authors may consider defining these terms so the general reader has a more concrete picture of the problem being solved here.

I've added a brief description of a sensor to the first sentence of the paper. I think our other recent updates help clarify what we mean by "optimization of sensor placement", but please let me know if you feel otherwise.
