### 2020-12-21 #2828 @pdebuyl
@briandesilva In basis_comparison notebook, I had to remove quiet=True from update_n_basis_modes function. Probably related to changes in the dependencies, can you check?

@jordanperr can you have another look at the code+paper and let us know whether you consider your comments addressed for publication?
