### 2020-12-22 9b5feb2
Add missing quiet option
