### 2020-12-22 #2828 @briandesilva
@pdebuyl, I just created a new release of the package—see if updating fixes the issue. I wasn't able to reproduce the problem on this version (0.3.1).
