### 2020-12-22 #2828 @pdebuyl
Fixed. I used the version from pip earlier, there might have been a mismatch between the pip installed and the git-downloaded examples.
