### 2021-02-09 #2828 @jordanperr
Well, looks like my Github notifications are not configured properly.. My sincere apologies for holding this up. Thanks @pdebuyl for sending me a nudge over email.

I have checked the remaining three items on the checklist. The authors have addressed all my comments and I **recommend acceptance for this paper**.
