### 2021-02-09 #2828 @pdebuyl
Thank you very much for completing the review @jordanperr

@briandesilva sorry for the delay. I will now proceed to the remaining editorial steps https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance
