### 2021-02-09 #2828 @pdebuyl
@briandesilva

Comments on the paper:
- The year in Candès is not showing properly
- The DOI recommendation by whedon for Narayanan seems correct, if so can you add it?

On the repo:
- Could you add a warning that the netcdf data example is a 200MB download? Maybe also check if the file exists, to avoid re-downloading the data?
- The examples run very well (no warning, nice results)!

You will need to archive your code on zenodo. https://joss.readthedocs.io/en/latest/submitting.html#the-review-process

Let me know about the doi and version of the archive.
