### 2021-02-09 #2828 @jordanperr
One final comment @briandesilva - I think some of the mathematical notation around line 56 is introduced without context. For example, I am not certain what the subscript y_i means. Since y is bold, it would appear to be a vector with length i, where i <= p. This happens again with C_*. What precisely does the subscript "star" mean in this context? Is it the same as s_* on line 65? I think one more pass of the mathematical notation to make sure each symbol is introduced and their meaning will be clear to broad audience.
