### 2021-02-13 f1493b5
Fix year and DOI of two references
