+++ docs/JOSS/paper.bib
@@ -15,7 +15,8 @@
   author={Narayanan, S Deepak and Patel, Zeel B and Agnihotri, Apoorv and Batra, Nipun},
   booktitle={Proceedings of the 18th Conference on Embedded Networked Sensor Systems},
   pages={653--654},
-  year={2020}
+  year={2020},
+  doi={10.1145/3384419.3430407},
 }
 
 @article{manohar2019optimized,
@@ -506,8 +507,9 @@
   Journal = {Communications in Pure and Applied Mathematics},
   Number = {1207--1223},
   Title = {Stable signal recovery from incomplete and inaccurate measurements},
-  Volume = {8},
-  Year = {59},
+  Volume = {59},
+  Number = {8},
+  Year = {2006},
   doi={10.1002/cpa.20124}}
 
 @article{Caselton1984spl,
