### 2021-02-14 6c9de67
Clarify mathematical notation and add more SSPOC details
