### 2021-02-14 #2828 @briandesilva
@jordanperr, I've done another pass over the math symbols, adding more detail where necessary. Thanks for the suggestion, and please let me know if you think there is anything in the paper that needs to be clarified further.

@pdebuyl, I've addressed the issues you raised and registered PySensors with Zenodo. The DOI is 10.5281/zenodo.4312270 and the archived version is v0.3.2.
