### 2021-02-15 #2828 @pdebuyl
Hi @briandesilva the doi for your archive is incorrect, it points to pysindy and not to PySensors
