### 2021-02-16 #2828 @briandesilva
Sorry about that @pdebuyl. The correct DOI is 10.5281/zenodo.4542530 and the version is v0.3.3.
