### 2021-02-16 #2828 @briandesilva
> Thanks @briandesilva . Can you edit the metadata of the zenodo archive to match the paper please? Title should be "PySensors: A Python Package for Sparse Sensor Placement" and authors should be as here, with ORCID properly set.

@pdebuyl, I've updated the archive as you asked.
