### 2021-02-17 #2828 @pdebuyl
For the eic: only orcid ids for authors seem missing from metadata.
