### 2021-02-18 #2828 @pdebuyl
Thanks @briandesilva . I don't think that we require them, I mentioned it as I believe that it brings value to the authors, to the readers, and to the journal :-)

The submission is in the hands of the Editors-in-chief in rotation.
