### 2021-02-27 #2828 @briandesilva
@arfon, I'm really sorry, but it was just pointed out to me that two of the author names were written in the wrong order in the paper. Is there any way to re-generate the pdf with the corrected order? I have corrected and pushed `paper.md`.
