### 2021-02-27 #2828 @arfon
Sure thing. Updated in https://github.com/openjournals/joss-papers/commit/8f2fc952f6c9bb5f9d32b1acd0a1f9ec2d96ad0a
